+++ superset/assets/src/visualizations/wordcloud/ReactWordCloud.js
@@ -1,0 +1,4 @@
+import reactify from '../../utils/reactify';
+import Component from './WordCloud';
+
+export default reactify(Component);
