+++ superset/assets/src/visualizations/wordcloud/adaptor.jsx
@@ -1,0 +1,5 @@
+import createAdaptor from '../../utils/createAdaptor';
+import Component from './ReactWordCloud';
+import transformProps from './transformProps';
+
+export default createAdaptor(Component, transformProps);
